+++ src/en/index.md
@@ -1,0 +1,10 @@
+---
+layout: home
+page: true
+---
+
+<script setup>
+import Home from '@theme/pages/Home.vue'
+</script>
+
+<Home />
